requirements.txt
mysql-connector-python==8.0.32
bcrypt==4.0.1
customtkinter==5.2.0
Pillow==10.0.0
matplotlib==3.7.1
fpdf==1.7.2
python-barcode==0.14.0
reportlab>=4.0.4
pandas>=2.0.0
openpyxl>=3.1.2
tkcalendar>=1.6.1
python-dotenv>=1.0.0 